+++ libraries/tools/kotlin-gradle-plugin/src/test/kotlin/org/jetbrains/kotlin/gradle/BaseGradleIT.kt
@@ -3,6 +3,7 @@ package org.jetbrains.kotlin.gradle
 import com.google.common.io.Files
 import com.intellij.openapi.util.io.FileUtil
 import org.gradle.api.logging.LogLevel
+import org.jetbrains.kotlin.gradle.plugin.ThreadTracker
 import org.jetbrains.kotlin.gradle.util.createGradleCommand
 import org.jetbrains.kotlin.gradle.util.runProcess
 import org.junit.After
@@ -66,7 +67,16 @@ abstract class BaseGradleIT {
     }
 
     // the second parameter is for using with ToolingAPI, that do not like --daemon/--no-daemon  options at all
-    data class BuildOptions(val withDaemon: Boolean = false, val daemonOptionSupported: Boolean = true)
+    data class BuildOptions(
+            val withDaemon: Boolean = false,
+            val daemonOptionSupported: Boolean = true,
+            val incremental: Boolean? = null,
+            /**
+             * @see [ThreadTracker]
+             */
+            val assertThreadLeaks: Boolean = true,
+            val androidHome: File? = null,
+            val androidGradlePluginVersion: String? = null)
 
     open inner class Project(val projectName: String, val wrapperVersion: String = "1.4", val minLogLevel: LogLevel = LogLevel.DEBUG) {
         open val resourcesRoot = File(resourcesRootFile, "testProject/$projectName")
@@ -88,8 +98,9 @@ abstract class BaseGradleIT {
         val compiledJavaSources: Iterable<File> by lazy { javaSourcesListRegex.findAll(output).asIterable().flatMap { it.groups[1]!!.value.split(" ").filter { it.endsWith(".java", ignoreCase = true) }.map { File(it).canonicalFile } } }
     }
 
-    fun Project.build(vararg tasks: String, options: BuildOptions = defaultBuildOptions(), check: CompiledProject.() -> Unit) {
-        val cmd = createBuildCommand(tasks, options)
+    fun Project.build(vararg params: String, options: BuildOptions = defaultBuildOptions(), check: CompiledProject.() -> Unit) {
+        val cmd = createBuildCommand(params, options)
+        val env = createEnvironmentVariablesMap(options)
 
         if (options.withDaemon) {
             prepareDaemon(wrapperVersion)
@@ -97,16 +108,12 @@ abstract class BaseGradleIT {
 
         println("<=== Test build: ${this.projectName} $cmd ===>")
 
-        runAndCheck(cmd, check)
-    }
-
-    private fun Project.runAndCheck(cmd: List<String>, check: CompiledProject.() -> Unit) {
         val projectDir = File(workingDir, projectName)
         if (!projectDir.exists()) {
             setupWorkingDir()
         }
 
-        val result = runProcess(cmd, projectDir)
+        val result = runProcess(cmd, projectDir, env)
         CompiledProject(this, result.output, result.exitCode).check()
     }
 
@@ -193,17 +200,30 @@ abstract class BaseGradleIT {
     private fun Project.createBuildCommand(params: Array<out String>, options: BuildOptions): List<String> =
             createGradleCommand(createGradleTailParameters(options, params))
 
-    protected fun Project.createGradleTailParameters(options: BuildOptions, params: Array<out String> = arrayOf()): List<String> =
-            params.asList() +
-                    listOf("-PpathToKotlinPlugin=" + File("local-repo").absolutePath,
-                            if (options.daemonOptionSupported)
-                                if (options.withDaemon) "--daemon"
-                                else "--no-daemon"
-                            else null,
-                            "--stacktrace",
-                            "--${minLogLevel.name.toLowerCase()}",
-                            "-Pkotlin.gradle.test=true")
-                            .filterNotNull()
+    private fun Project.createGradleTailParameters(options: BuildOptions, params: Array<out String> = arrayOf()): List<String> =
+            params.toMutableList().apply {
+                add("--stacktrace")
+                add("--${minLogLevel.name.toLowerCase()}")
+                if (options.daemonOptionSupported) {
+                    add(if (options.withDaemon) "--daemon" else "--no-daemon")
+                }
+
+                add("-PpathToKotlinPlugin=" + File("local-repo").absolutePath)
+                if (options.assertThreadLeaks) {
+                    add("-P${ThreadTracker.ASSERT_THREAD_LEAKS_PROPERTY}=true")
+                }
+                options.incremental?.let { add("-Pkotlin.incremental=$it") }
+                options.androidGradlePluginVersion?.let { add("-PandroidToolsVersion=$it")}
+            }
+
+    private fun Project.createEnvironmentVariablesMap(options: BuildOptions): Map<String, String> =
+            hashMapOf<String, String>().apply {
+                val sdkDir = options.androidHome
+                if (sdkDir != null) {
+                    sdkDir.parentFile.mkdirs()
+                    put("ANDROID_HOME", sdkDir.canonicalPath)
+                }
+            }
 
     private fun String.normalize() = this.lineSequence().joinToString(SYSTEM_LINE_SEPARATOR)